+++ NEWS.md
@@ -16,6 +16,7 @@ v0.4.0 (coming soon)
 * Add new test matrices:
 
 	- `golub` see http://blogs.mathworks.com/cleve/2015/08/24/golub-matrices-deceptively-ill-conditioned/
+    - `companion`: Companion matrix.
 
 v0.3.5 
 ----------